requirements.txt
flask
psycopg2-binary
pandas
os
